refactor large editor component
