+++ src/tabs/editor/components/single-company/SingleCompanyView.tsx
@@ -1,73 +1,19 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import {
-  AlertCircle,
-  ArrowLeft,
-  CheckCircle2,
-  Minus,
-} from "lucide-react";
+import { useCallback, useEffect, useRef, useState } from "react";
+import { ArrowLeft } from "lucide-react";
 import { useNavigate, useParams } from "react-router-dom";
 import { useI18n } from "@/contexts/I18nContext";
 import { Button } from "@/ui/button";
 import { LoadingSpinner } from "@/ui/loading-spinner";
-import { listCompanies, getCompany } from "../../lib/companies-api";
-import { fetchTagOptions } from "../../lib/tag-options-api";
-import type {
-  GarboCompanyDetail,
-  GarboCompanyListItem,
-  TagOption,
-} from "../../lib/types";
-import type { VerificationState } from "../../lib/verification";
-import { getCompanyVerificationOverview } from "../../lib/verification";
-import { SingleSelectDropdown } from "@/ui/single-select-dropdown";
-import { MultiSelectDropdown } from "@/ui/multi-select-dropdown";
+import { getCompany } from "../../lib/companies-api";
+import type { GarboCompanyDetail } from "../../lib/types";
 import { CompanyEditDetail } from "./CompanyEditDetail";
-import { DataTable, DataTableBody, DataTableHead, DataTableShell } from "@/ui/data-table";
-import { NO_TAGS_FILTER_OPTION } from "../../lib/types";
+import { EDITOR_INDEX_PATH } from "../../lib/editor-routes";
+import { useSingleCompanyOverviewList } from "../../hooks/useSingleCompanyOverviewList";
 import {
-  buildTagLabelBySlug,
-  companyMatchesTagFilter,
-  companyPassesExcludeTagFilter,
-} from "../../lib/editor-tag-and-payload-utils";
-import { SearchAndFiltersCard } from "@/ui/search-and-filters-card";
-import { ReportingPeriodQuickEditModal } from "./ReportingPeriodQuickEditModal";
-import { displayBaseYear } from "../../lib/company-edit-utils";
-import { getPeriodYear } from "../../lib/reporting-period-ui";
-import {
-  EDITOR_INDEX_PATH,
-  editorCompanyPath,
-} from "../../lib/editor-routes";
-
-function companyHasPeriodsInYears(
-  company: GarboCompanyListItem,
-  years: string[]
-): boolean {
-  if (!years.length) return true;
-  // AND: must have all selected years
-  return years.every((y) =>
-    (company.reportingPeriods ?? []).some((p) => getPeriodYear(p) === y)
-  );
-}
-
-function companyMatchesSearch(
-  company: GarboCompanyListItem,
-  query: string
-): boolean {
-  const q = query.trim().toLowerCase();
-  if (!q) return true;
-  const name = (company.name ?? "").toLowerCase();
-  const id = (company.wikidataId ?? "").toLowerCase();
-  return name.includes(q) || id.includes(q);
-}
-
-function StatusIcon({ state }: { state: VerificationState }) {
-  if (state === "verified")
-    return <CheckCircle2 className="w-4 h-4 text-green-03" />;
-  if (state === "unverified")
-    return <AlertCircle className="w-4 h-4 text-orange-03" />;
-  return <Minus className="w-4 h-4 text-gray-03" />;
-}
-
-type CompanySortId = "name-asc" | "name-desc" | "id-asc" | "id-desc";
+  SingleCompanyOverviewFilters,
+  SingleCompanyOverviewListAfterSlot,
+} from "./SingleCompanyOverviewFilters";
+import { SingleCompanyOverviewTable } from "./SingleCompanyOverviewTable";
 
 export function SingleCompanyView() {
   const { t } = useI18n();
@@ -75,61 +21,21 @@ export function SingleCompanyView() {
   const { companyId: companyIdParam } = useParams<{ companyId?: string }>();
   const routeCompanyId = companyIdParam ? decodeURIComponent(companyIdParam) : null;
 
-  const dash = t("common.placeholderDash");
-  const [companyList, setCompanyList] = useState<GarboCompanyListItem[]>([]);
-  const [tagOptions, setTagOptions] = useState<TagOption[]>([]);
-  const [loadingList, setLoadingList] = useState(true);
-  const [listError, setListError] = useState<string | null>(null);
-
-  const [searchQuery, setSearchQuery] = useState("");
-  const [filterTags, setFilterTags] = useState<string[]>([]);
-  const [excludeFilterTags, setExcludeFilterTags] = useState<string[]>([]);
-  const [filterYears, setFilterYears] = useState<string[]>([]);
-  const [filterSector, setFilterSector] = useState<string>("");
-  const [filterHasUnverifiedEmissions, setFilterHasUnverifiedEmissions] = useState(false);
-  const [filterHasUnverifiedData, setFilterHasUnverifiedData] = useState(false);
-  const [filtersOpen, setFiltersOpen] = useState(true);
-  const [companySort, setCompanySort] = useState<CompanySortId>("name-asc");
+  const overviewList = useSingleCompanyOverviewList();
 
   const [detail, setDetail] = useState<GarboCompanyDetail | null>(null);
   const [loadingDetail, setLoadingDetail] = useState(false);
   const [detailError, setDetailError] = useState<string | null>(null);
-  const [quickEdit, setQuickEdit] = useState<{ companyId: string; year: string } | null>(null);
+  const [quickEdit, setQuickEdit] = useState<{ companyId: string; year: string } | null>(
+    null,
+  );
 
-  // Guards against stale async operations (e.g. save finishing after navigation).
   const activeCompanyIdRef = useRef<string | null>(routeCompanyId);
   const detailRefreshTokenRef = useRef(0);
-
-  const tagLabelBySlug = useMemo(() => buildTagLabelBySlug(tagOptions), [tagOptions]);
 
   useEffect(() => {
     activeCompanyIdRef.current = routeCompanyId;
   }, [routeCompanyId]);
-
-  useEffect(() => {
-    let cancelled = false;
-    setLoadingList(true);
-    setListError(null);
-    Promise.all([listCompanies(), fetchTagOptions()])
-      .then(([list, tags]) => {
-        if (!cancelled) {
-          setCompanyList(list);
-          setTagOptions(tags);
-        }
-      })
-      .catch((e) => {
-        if (!cancelled) {
-          setListError(e instanceof Error ? e.message : String(e));
-          setCompanyList([]);
-        }
-      })
-      .finally(() => {
-        if (!cancelled) setLoadingList(false);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, []);
 
   useEffect(() => {
     if (!routeCompanyId) {
@@ -158,102 +64,6 @@ export function SingleCompanyView() {
       cancelled = true;
     };
   }, [routeCompanyId]);
-
-  /** All unique reporting period years (from end date) in the companies list, newest first. */
-  const years = useMemo(() => {
-    const set = new Set<string>();
-    companyList.forEach((c) => {
-      c.reportingPeriods?.forEach((p) => {
-        const y = getPeriodYear(p);
-        if (y) set.add(y);
-      });
-    });
-    return Array.from(set).sort((a, b) => b.localeCompare(a));
-  }, [companyList]);
-
-  const sectors = useMemo(() => {
-    const set = new Set<string>();
-    companyList.forEach((c) => {
-      const code = c.industry?.subIndustryCode;
-      if (code) set.add(code);
-    });
-    return Array.from(set).sort();
-  }, [companyList]);
-
-  const companyOverviewById = useMemo(() => {
-    const map = new Map<string, ReturnType<typeof getCompanyVerificationOverview>>();
-    companyList.forEach((c) => {
-      map.set(c.wikidataId, getCompanyVerificationOverview(c));
-    });
-    return map;
-  }, [companyList]);
-
-  const filteredCompanies = useMemo(() => {
-    return companyList.filter((c) => {
-      const overview = companyOverviewById.get(c.wikidataId);
-      if (!companyMatchesSearch(c, searchQuery)) return false;
-      if (!companyMatchesTagFilter(c.tags, filterTags)) return false;
-      if (!companyPassesExcludeTagFilter(c.tags, excludeFilterTags)) return false;
-      if (filterYears.length && !companyHasPeriodsInYears(c, filterYears))
-        return false;
-      if (
-        filterSector &&
-        (c.industry?.subIndustryCode ?? "") !== filterSector
-      )
-        return false;
-      if (filterHasUnverifiedEmissions && !overview?.hasUnverifiedEmissions)
-        return false;
-      if (filterHasUnverifiedData && !overview?.hasUnverifiedData) return false;
-      return true;
-    });
-  }, [
-    companyList,
-    companyOverviewById,
-    searchQuery,
-    filterTags,
-    excludeFilterTags,
-    filterYears,
-    filterSector,
-    filterHasUnverifiedEmissions,
-    filterHasUnverifiedData,
-  ]);
-
-  const sortedCompanies = useMemo(() => {
-    const rows = [...filteredCompanies];
-    const byName = (a: GarboCompanyListItem, b: GarboCompanyListItem) =>
-      (a.name ?? "").localeCompare(b.name ?? "", undefined, {
-        sensitivity: "base",
-      });
-    const byId = (a: GarboCompanyListItem, b: GarboCompanyListItem) =>
-      (a.wikidataId ?? "").localeCompare(b.wikidataId ?? "");
-    switch (companySort) {
-      case "name-desc":
-        rows.sort((a, b) => -byName(a, b));
-        break;
-      case "id-asc":
-        rows.sort(byId);
-        break;
-      case "id-desc":
-        rows.sort((a, b) => -byId(a, b));
-        break;
-      default:
-        rows.sort(byName);
-    }
-    return rows;
-  }, [filteredCompanies, companySort]);
-
-  const filterPeriodStats = useMemo(() => {
-    let totalPeriods = 0;
-    let verifiedEmissionsPeriods = 0;
-    for (const c of filteredCompanies) {
-      const overview = companyOverviewById.get(c.wikidataId);
-      for (const p of overview?.perYear ?? []) {
-        totalPeriods += 1;
-        if (p.emissions === "verified") verifiedEmissionsPeriods += 1;
-      }
-    }
-    return { totalPeriods, verifiedEmissionsPeriods };
-  }, [filteredCompanies, companyOverviewById]);
 
   const goBack = useCallback(() => {
     navigate(EDITOR_INDEX_PATH);
@@ -313,14 +123,12 @@ export function SingleCompanyView() {
         </div>
         <CompanyEditDetail
           company={detail}
-          tagOptions={tagOptions}
+          tagOptions={overviewList.tagOptions}
           onSaved={() => {
             const companyIdAtSave = routeCompanyId;
 
-            // Refresh the list (safe regardless of which company is currently open).
-            listCompanies().then(setCompanyList);
+            void overviewList.refreshCompanyList();
 
-            // Refresh the currently-saved company detail, but ignore stale results.
             if (!companyIdAtSave) return;
 
             const token = ++detailRefreshTokenRef.current;
@@ -330,347 +138,28 @@ export function SingleCompanyView() {
                 if (detailRefreshTokenRef.current !== token) return;
                 setDetail(company);
               })
-              .catch(() => {
-                // Ignore: detail fetch errors are handled by the route-level effect.
-              });
+              .catch(() => {});
           }}
         />
       </div>
     );
   }
 
+  const dash = t("common.placeholderDash");
+
   return (
     <div className="space-y-4">
-      <SearchAndFiltersCard
-        title={t("editor.singleCompanyView.searchAndFilters")}
-        open={filtersOpen}
-        onOpenChange={setFiltersOpen}
-        after={
-          <>
-            {listError && !loadingList && (
-              <div className="rounded-lg border border-gray-03 bg-gray-05/80 p-4">
-                <p className="text-gray-01 font-medium">
-                  {t("editor.singleCompanyView.loadError")}
-                </p>
-                <p className="text-sm text-gray-02 mt-1">{listError}</p>
-              </div>
-            )}
+      <SingleCompanyOverviewFilters
+        list={overviewList}
+        afterSlot={<SingleCompanyOverviewListAfterSlot list={overviewList} />}
+      />
 
-            {loadingList && (
-              <div className="flex justify-center py-12">
-                <LoadingSpinner label={t("editor.companies.loading")} />
-              </div>
-            )}
-
-            {!loadingList && filteredCompanies.length === 0 && (
-              <div className="py-8 text-center text-gray-02 text-sm">
-                {t("editor.singleCompanyView.noCompaniesMatch")}
-              </div>
-            )}
-          </>
-        }
-      >
-        <div>
-          <label className="block text-xs font-medium text-gray-02 mb-1">
-            {t("editor.singleCompanyView.searchByNameOrId")}
-          </label>
-          <input
-            type="text"
-            value={searchQuery}
-            onChange={(e) => setSearchQuery(e.target.value)}
-            placeholder={t("editor.singleCompanyView.searchPlaceholder")}
-            className="w-full max-w-md px-3 py-2 rounded-lg border border-gray-03 bg-gray-05 text-gray-01 placeholder:text-gray-03 focus:outline-none focus:ring-2 focus:ring-blue-03"
-          />
-        </div>
-        <div className="flex flex-wrap gap-4">
-          <div>
-            <label className="block text-xs font-medium text-gray-02 mb-1">
-              {t("editor.companies.tag")}
-            </label>
-            <MultiSelectDropdown
-              options={[NO_TAGS_FILTER_OPTION, ...tagOptions.map((o) => o.slug)]}
-              selectedIds={filterTags}
-              onChange={setFilterTags}
-              triggerLabel={t("editor.companies.tags")}
-              getOptionLabel={(slug) =>
-                slug === NO_TAGS_FILTER_OPTION
-                  ? t("editor.companies.noTags")
-                  : (tagLabelBySlug[slug] ?? slug)
-              }
-              emptyLabel={t("editor.companies.allTags")}
-              triggerClassName="min-w-[140px] !h-8 !text-xs px-3"
-            />
-          </div>
-          <div>
-            <label className="block text-xs font-medium text-gray-02 mb-1">
-              {t("editor.singleCompanyView.excludeTagsLabel")}
-            </label>
-            <MultiSelectDropdown
-              options={tagOptions.map((o) => o.slug)}
-              selectedIds={excludeFilterTags}
-              onChange={setExcludeFilterTags}
-              triggerLabel={t("editor.singleCompanyView.excludeTagsTrigger")}
-              getOptionLabel={(slug) => tagLabelBySlug[slug] ?? slug}
-              emptyLabel={t("editor.singleCompanyView.excludeTagsEmpty")}
-              triggerClassName="min-w-[140px] !h-8 !text-xs px-3"
-            />
-          </div>
-          <div>
-            <label className="block text-xs font-medium text-gray-02 mb-1">
-              {t("editor.companies.year")}
-            </label>
-            <MultiSelectDropdown
-              options={years}
-              selectedIds={filterYears}
-              onChange={setFilterYears}
-              triggerLabel={t("editor.companies.year")}
-              emptyLabel={t("editor.companies.allYears")}
-              triggerClassName="min-w-[120px] !h-8 !text-xs px-3"
-            />
-          </div>
-          {sectors.length > 0 && (
-            <div>
-              <label className="block text-xs font-medium text-gray-02 mb-1">
-                {t("editor.singleCompanyView.sector")}
-              </label>
-              <SingleSelectDropdown
-                options={["", ...sectors]}
-                value={filterSector}
-                onChange={setFilterSector}
-                placeholder={t("editor.singleCompanyView.allSectors")}
-                getOptionLabel={(v) => (v ? v : t("editor.singleCompanyView.allSectors"))}
-                triggerClassName="min-w-[140px] !h-8 !text-xs px-3"
-              />
-            </div>
-          )}
-          <div className="flex flex-wrap items-end gap-4">
-            <label className="flex items-center gap-2 cursor-pointer text-gray-01 text-xs">
-              <input
-                type="checkbox"
-                checked={filterHasUnverifiedEmissions}
-                onChange={(e) => setFilterHasUnverifiedEmissions(e.target.checked)}
-                className="rounded border-gray-03"
-              />
-              {t("editor.singleCompanyView.filterHasUnverifiedEmissions")}
-            </label>
-            <label className="flex items-center gap-2 cursor-pointer text-gray-01 text-xs">
-              <input
-                type="checkbox"
-                checked={filterHasUnverifiedData}
-                onChange={(e) => setFilterHasUnverifiedData(e.target.checked)}
-                className="rounded border-gray-03"
-              />
-              {t("editor.singleCompanyView.filterHasUnverifiedData")}
-            </label>
-          </div>
-        </div>
-        {!loadingList && filteredCompanies.length > 0 && (
-          <p className="text-xs text-gray-02 mt-3 pt-3 border-t border-gray-03/50">
-            {t("editor.singleCompanyView.filterPeriodValidationSummary", {
-              verified: filterPeriodStats.verifiedEmissionsPeriods,
-              total: filterPeriodStats.totalPeriods,
-            })}
-          </p>
-        )}
-      </SearchAndFiltersCard>
-
-      {!loadingList && filteredCompanies.length > 0 && (
-        <div className="flex flex-col gap-3 sm:flex-row sm:flex-wrap sm:items-center sm:justify-between rounded-lg border border-gray-03 bg-gray-05/60 px-4 py-3">
-          <p className="text-sm text-gray-01">
-            {t("editor.singleCompanyView.tableStatsLine", {
-              shown: filteredCompanies.length,
-              total: companyList.length,
-              verifiedPeriods: filterPeriodStats.verifiedEmissionsPeriods,
-              periodTotal: filterPeriodStats.totalPeriods,
-            })}
-          </p>
-          <div className="flex flex-wrap items-center gap-2">
-            <span className="text-xs font-medium text-gray-02">
-              {t("editor.singleCompanyView.sortLabel")}
-            </span>
-            <SingleSelectDropdown
-              options={
-                [
-                  "name-asc",
-                  "name-desc",
-                  "id-asc",
-                  "id-desc",
-                ] satisfies CompanySortId[]
-              }
-              value={companySort}
-              onChange={(v) => setCompanySort(v as CompanySortId)}
-              getOptionLabel={(id) => {
-                switch (id) {
-                  case "name-desc":
-                    return t("editor.singleCompanyView.sortNameDesc");
-                  case "id-asc":
-                    return t("editor.singleCompanyView.sortIdAsc");
-                  case "id-desc":
-                    return t("editor.singleCompanyView.sortIdDesc");
-                  default:
-                    return t("editor.singleCompanyView.sortNameAsc");
-                }
-              }}
-              placeholder={t("editor.singleCompanyView.sortNameAsc")}
-              triggerClassName="min-w-[200px] !h-8 !text-xs px-3"
-            />
-          </div>
-        </div>
-      )}
-
-      {!loadingList && filteredCompanies.length > 0 && (
-        <DataTableShell>
-          {quickEdit && (
-            <ReportingPeriodQuickEditModal
-              open={true}
-              onOpenChange={(open) => {
-                if (!open) setQuickEdit(null);
-              }}
-              company={companyList.find((x) => x.wikidataId === quickEdit.companyId)!}
-              year={quickEdit.year}
-              onSaved={() => {
-                listCompanies().then(setCompanyList);
-              }}
-            />
-          )}
-          <DataTable>
-            <DataTableHead>
-              <tr>
-                <th className="px-4 py-3 font-medium w-[22%]">
-                  {t("editor.companies.company")}
-                </th>
-                <th className="px-4 py-3 font-medium">
-                  <div className="leading-tight">
-                    <div>{t("editor.singleCompanyView.table.baseYear")}</div>
-                    <div className="text-[10px] text-gray-01 normal-case tracking-normal">
-                      {t("editor.singleCompanyView.table.allYears")}
-                    </div>
-                  </div>
-                </th>
-                <th className="px-4 py-3 font-medium">
-                  <div className="leading-tight">
-                    <div>{t("editor.singleCompanyView.table.emissions")}</div>
-                    <div className="text-[10px] text-gray-01 normal-case tracking-normal">
-                      {t("editor.singleCompanyView.table.allYears")}
-                    </div>
-                  </div>
-                </th>
-                <th className="px-4 py-3 font-medium">
-                  <div className="leading-tight">
-                    <div>{t("editor.singleCompanyView.table.economy")}</div>
-                    <div className="text-[10px] text-gray-01 normal-case tracking-normal">
-                      {t("editor.singleCompanyView.table.allYears")}
-                    </div>
-                  </div>
-                </th>
-                <th className="px-4 py-3 font-medium">
-                  {t("editor.companies.year")}
-                </th>
-                <th className="px-4 py-3 font-medium">
-                  {t("editor.singleCompanyView.sections.industry")}
-                </th>
-                <th className="px-4 py-3 font-medium w-[18%]">
-                  {t("editor.companies.tags")}
-                </th>
-              </tr>
-            </DataTableHead>
-            <DataTableBody>
-              {sortedCompanies.map((c) => {
-                const overview = companyOverviewById.get(c.wikidataId);
-                return (
-                  <tr
-                    key={c.wikidataId}
-                    onClick={() => navigate(editorCompanyPath(c.wikidataId))}
-                    className="hover:bg-gray-04/50 cursor-pointer text-gray-01 align-top"
-                  >
-                    <td className="px-4 py-3 font-medium">
-                      <div className="flex flex-col">
-                        <span>{c.name}</span>
-                        <span className="text-xs text-gray-02">{c.wikidataId}</span>
-                      </div>
-                    </td>
-                    <td className="px-4 py-3 text-gray-02">
-                      <div className="flex items-center gap-2">
-                        <StatusIcon state={overview?.baseYear ?? "none"} />
-                        <span>{displayBaseYear(c.baseYear, dash)}</span>
-                      </div>
-                    </td>
-                    <td className="px-4 py-3">
-                      <StatusIcon state={overview?.emissions ?? "none"} />
-                    </td>
-                    <td className="px-4 py-3">
-                      <StatusIcon state={overview?.economy ?? "none"} />
-                    </td>
-                    <td className="px-4 py-3 text-gray-02">
-                      {overview?.perYear?.length ? (
-                        <div className="flex flex-wrap gap-2">
-                          {overview.perYear.map((p) => (
-                            <button
-                              key={p.year}
-                              type="button"
-                              onClick={(e) => {
-                                e.stopPropagation();
-                                setQuickEdit({ companyId: c.wikidataId, year: p.year });
-                              }}
-                              className="inline-flex items-center gap-2 rounded-full border border-gray-03 px-2 py-1 text-xs text-gray-01 bg-gray-05 hover:bg-gray-03/40"
-                              title={t("editor.periodEditor.quickEditYearTitle", { year: p.year })}
-                            >
-                              <span className="font-semibold">{p.year}</span>
-                              <span className="inline-flex items-center gap-1">
-                                <span className="text-[10px] text-gray-03">E</span>
-                                <StatusIcon state={p.emissions} />
-                              </span>
-                              <span className="inline-flex items-center gap-1">
-                                <span className="text-[10px] text-gray-03">$</span>
-                                <StatusIcon state={p.economy} />
-                              </span>
-                            </button>
-                          ))}
-                        </div>
-                      ) : (
-                        dash
-                      )}
-                    </td>
-                    <td className="px-4 py-3 text-gray-02">
-                      <div className="flex items-center gap-2">
-                        <StatusIcon state={overview?.industry ?? "none"} />
-                        <span>{c.industry?.subIndustryCode ?? dash}</span>
-                      </div>
-                    </td>
-                    <td className="px-4 py-3 text-gray-02">
-                      {c.tags?.length ? (
-                        <div className="flex flex-wrap gap-1.5">
-                          {c.tags.slice(0, 4).map((slug) => (
-                            <span
-                              key={slug}
-                              className="inline-flex items-center px-2 py-0.5 rounded-full text-[11px] font-medium bg-gray-03/40 text-gray-01 border border-gray-03"
-                            >
-                              {tagLabelBySlug[slug] ?? slug}
-                            </span>
-                          ))}
-                          {c.tags.length > 4 && (
-                            <span className="text-[11px] text-gray-02">
-                              +{c.tags.length - 4}
-                            </span>
-                          )}
-                        </div>
-                      ) : (
-                        dash
-                      )}
-                    </td>
-                  </tr>
-                );
-              })}
-            </DataTableBody>
-          </DataTable>
-          <p className="px-4 py-2 text-xs text-gray-01/90 border-t border-gray-03/50">
-            {t("editor.singleCompanyView.showingCount", {
-              count: filteredCompanies.length,
-              total: companyList.length,
-            })}
-          </p>
-        </DataTableShell>
-      )}
+      <SingleCompanyOverviewTable
+        list={overviewList}
+        dash={dash}
+        quickEdit={quickEdit}
+        onQuickEditChange={setQuickEdit}
+      />
     </div>
   );
 }
